Continue after the warning about missing 'units'
When regridding a variable that does not have the "units" attribute PISM used to
1) warn that the "units" attribute is missing and then
2) stop with an error message saying something like "cannot convert from 'm' to ''".

This does not make sense: we should either warn or stop with an error message.

Now PISM warns that the "units" attribute is missing and assumes that units in the file
match PISM's (internal) units but *does not* stop.