Removed PETSC_EXECUTABLE_RUNS check from the FindPETSc test. This was a check that tested PETSc is functional by compiling and trying to run the program :
    static const char help[] = "PETSc test program.";
    #include <petscts.h>
    int main(int argc,char *argv[]) {
      PetscErrorCode ierr;
      TS ts;
        printf("Starting\n");
      ierr = PetscInitialize(&argc,&argv,0,help);CHKERRQ(ierr);
      ierr = TSCreate(PETSC_COMM_WORLD,&ts);CHKERRQ(ierr);
      ierr = TSSetFromOptions(ts);CHKERRQ(ierr);
      ierr = TSDestroy(&ts);CHKERRQ(ierr);
      ierr = PetscFinalize();CHKERRQ(ierr);
        printf("Finished\n");
      return 0;
    }

It was failing on the discover supercomputer with SLURM and Intel MPI 20 with the error:

    *Abort(2140047) on node 0 (rank 0 in comm 0): Fatal error in
    PMPI_Init_thread: Other MPI error, error stack:*
    *MPIR_Init_thread(703)........: *
    *MPID_Init(958)...............: *
    *MPIDI_OFI_mpi_init_hook(1334): *
    *MPIDU_bc_table_create(444)...: *

Building the test program by hand and running directly on the discover login node (i.e. not in SLURM) resulted in:
   a) ./a.out is run directly: the same error
   b) If run with mpirun ./a.out: it hung

I then tried running the (now compiled) program within slurm, using:

    sbatch --constraint=hasw --account=xxxx --ntasks=2 --time=0:01:01 --qos=debug <hello.sh

    Where hello.sh is:

    #!/bin/bash
    #

    mpirun ./a.out

Everything worked successfully.  Conclusion: the test program FindPETSc() wants to run, which requires MPI, can only run properly within SLURM, i.e. NOT within a CMake setup (compiling on a login node).  Since this is an old version of PISM, the obvious hack is to comment out the test; as I have shown manually that the tested behavior works.