Hide implementation details
... by using forward declarations in headers.

Any program using PISM as a library has to be able to find header files included in PISM's
public headers.

This may be OK in the case of GSL, FFTW, UDUNITS, and NetCDF (these are usually installed
in standard locations), but including PETSc's headers in PISM's public headers can lead to
more difficulties.

This commit reduces the number of library headers included in PISM's headers.