Support PROJ 6.x properly
Now that we support proj="EPSG:XXXX" to initialize projection info we have to be careful
about this PROJ quirk: "+init=epsg:XXXX" creates a *transformation* and "epsg:XXXX"
creates a coordinate reference system.

I think we just need to create a transformation from a given CRS to lat,lon (EPGS:4326)
and use that.