Fix up an old regression test
This regression test is supposed to check that the difference between the basal
temperature in the ice and the temperature at the top of the bedrock decreases as you
refine in time.

At some point in time the test broke: after some change PISM computes higher
diffusivities, making the -max_dt option in the test ineffective (with adaptive time
steppint choosing time steps of ~50 years -max_dt 100 does nothing).

This change was not caught because the test incorrectly used ">" instead of ">=" to decide
if it failed.

This commit cleans up the script, removing its dependence on NCO and fixes the bug
described above.