Compute interface fluxes and *then* limit them
This commit removes a premature optimization: now we compute interface fluxes in two steps

1) Compute fluxes
2) Limit them

Also, it should not be necessary to use the velocity BC mask. Before this commit we used
the velocity BC mask in a way that allowed one to think about BC locations as locations
where ice *flux* is prescribed, but this was not used. Now it is a little harder to
predict the flux resulting from prescribing ice velocity and thickness, but the code is
simpler and more extensible.