Add a surface slope limiter
This limiter reduces (and in some cases eliminates) unreasonable velocity spikes near
steep margins.

The idea comes from CISM's Blatter solver, but the implementation is different: we use
maximum cliff height dh (configuration parameter stress_balance.blatter.max_cliff_height,
default of 200 meters) instead of the unitless maximum slope. This way the limit is
automatically adjusted with the grid resolution. (The maximum slope is set to
dh / max(dx, dy).)

The default value chosen appears to be effective *and* is close to the one suggested in
"Upper and lower limits on the stability of calving glaciers from the yield strength
envelope of ice" by Bassis and Walker (2011) (doi:10.1098/rspa.2011.0422).

PS: CISM has the same issue but uses an "upstream" finite difference approximation
instead of limiting the slope. It is worth noting that due to the shape of the margin
this approach *also* has the effect of reducing the slope.