Fix the name of a configuration parameter (ocean.th.clip_salinity)
This parameter used to be called ocean.three_equation_model_clip_salinity. A recent commit
renamed it to ocean.th.three_equation_model_clip_salinity, but that is
redundant ("ocean.th" is the "three equation model").

Also: update the reference to this parameter in the manual (note: this inconsistency was
caught by PISM's CI setup).