Fully implement M* scheme, including upwinding.
This should be the full M* scheme for the isothermal SIA, i.e. this is the
explicit time-stepping version of the scheme in Bueler 2016.  However,
it is only implemented for the isothermal SIA, and it is badly implemented
because of code duplication and known-to-be slow choices.  (See FIXMEs in
W_mstar() and q_mstar().)  No non-flat bed testing has yet been done, but it
compiles and produces the same numbers as the previous M* implementation on
flat-bed Test B.