Support checkpointing the HTCondor way
HTCondor supports  "self-checkpointing" applications that

1) stop with a specific exit code after saving a checkpoint file
2) can continue from a checkpoint file *with the same command line arguments* if this file
   is present
3) can re-start relatively quickly

The point of *stopping* after saving a checkpoint file (item 1) is to allow HTCondor to
transfer the model state from the machine executing PISM to the submit host so that we can
continue the run on a (possibly) different machine.

Stopping with a specific exit code is needed to be able to recognize runs that need to
continue (as opposed to runs that completed).

PISM itself does not need to check if a checkpoint file is present (item 2): this can be
done in a shell script that runs PISM.

PISM's "backup" mechanism is *almost* what we need here, so I renamed
`output.backup_interval` to `output.checkpoint.interval`, `output.backup_size` to
`output.checkpoint.size` and added

- `output.checkpoint.exit`: flag that should be set to `true` to stop after saving a
  checkpoint file

- `output.checkpoint.exit_code`: the exit code to use (the default is 85, corresponding to
  ERESTART on Linux (suggested by the HTCondor manual))

- `output.checkpoint.file` name of the checkpoint file