Use `realpath()` to resolve relative file names
Now we use POSIX `realpath()` to get absolute file names and record them in the
configuration database.

This affects configuration parameters that have names ending in ".file" (and only these).

This will make it easier to re-run a simulation that produced a particular output file
because canonical file names of all input files are now recorded in each output file.