Remove a misguided energy conservation attempt (AKA Hfrozen)
This commit removes an energy conservation attempt that seemed like a good idea at the
time but turned out to do more harm than good.

See the comment in the diff of this commit for details.

Here's the context.

To preserve continuity of temperature in the column of ice and bedrock *at the base of the
ice* `EnthalpyModel::update_impl()` ensures that if the till water thickness is not zero
then the basal enthalpy cannot be below the CTS enthalpy (in other words, ice at the base
has to be temperate).

In other words, if (E_base < E_CTS_base) *after* the time step of the energy model but
before post-processing, then E_base is set to E_CTS_base, i.e. E_base is increased. This
correction creates *internal energy* out of nothing.

Prior to this commit PISM attempted to remove the equivalent amount of energy from the
basal till water layer, freezing some of the water by modifying the basal melt rate
accordingly.

Unfortunately (both because of adaptive time stepping and because of asynchronous coupling
between energy conservation and mass transport models when `time_stepping.skip.enabled` is
set) the enthalpy model does not know if the basal till storage has enough water to freeze
and cannot easily limit the magnitude of this correction. As a result in some
hard-to-reproduce settings this correction ends up being large enough to negatively affect
mass conservation, creating *mass* out of nothing (freezing more water than available).
Occasionally this mechanism created five-kilometer-tall columns of ice -- enough to see a
spike it the reported total ice mass (on coarse Greenland-wide grids, anyway) and to cause
a crash with the error message "ice thickness is too high."

It *is* possible to keep this correction and avoid *some* mass conservation issues, but we
would have to do the following.

- Re-order operations performed in `IceModel::step()` so that `IceModel::hydrology_step()`
  is called before the mass continuity step.

- Modify hydrology models so that they provide the "effective" melt rate for a given time
  step. This quantity would be the same as the energy-balance-determined basal melt rate
  where the melt rate is zero or positive while freeze-on (negative melt rate) would be
  reduced to avoid freezing more water than available.

- Use this "effective basal melt rate" to perform the mass continuity step.

Note, though, that there is a similar issue with *positive* melt rates: it is possible
that in some cases a hydrology model may receive more water than is available in the
column of ice above: the mass continuity code "caps" the melt rate to preserve
non-negativity of ice thickness but hydrology models do not know about that.

However, even if I implement this properly, positive effects of this correction are likely
to be drowned out by other energy-conservation errors. For example, the "bulge
limiter" (see `EnthalpyModel::update_impl()`) also creates energy out of nothing but we do
not attempt to remove energy from elsewhere in the system to restore the balance.