Extrapolate sliding (SSA) velocity at ice margins
This may help improve the initial guess of ice viscosity in the outer (Picard) iteration
in SSAFD during the next time step. Without this (and with CFBC) ice velocity in ice-free
areas is set to (0.0, 0.0) and FD approximations of partial derivatives of (u, v) are of
poor quality *if* the ice advances into a previously-ice-free cell. Estimates of u_x, v_y,
etc affect the quality of the estimate of the ice viscosity.

Set stress_balance.ssa.fd.extrapolate_at_margins to false to disable.