Add CMake option Pism_PKG_CONFIG_STATIC (default: on) (see #529)
Normally it is enough to set "-L${petsc_prefix}/lib -lpetsc" to be able to link to PETSc:
libpetsc.so should have names and locations of its dependencies recorded in its metadata.

This mechanism allowing a linker to automatically discover indirect library dependencies
appears to be broken on some systems, requiring explicitly listing both direct and
indirect dependencies.

To try to avoid this issue PISM's CMake scripts run pkg-config with the flag "--static" to
get more complete lists of linker flags (usually used for static linking since static
libraries are just collections of object files and don't contain metadata).

This may lead to "overlinking", i.e. listing some libraries as *direct* dependencies of
libpism.so even though PISM itself does not use them. They would have to be loaded (either
as direct dependencies or not) anyway, so this appears to be the only side effect. I think
we can live with that.

Set this option to "off" to use pkg-config without the "--static" flag.

Note that we had to require CMake >= 3.16 for this to work properly (see
https://github.com/Kitware/CMake/commit/28cb86d7962b2a9b8e49dc62bccf6187761524d9).