Climate Index Implementation
Implementation of a Climate Climate Index timeseries for glacial/interglacial simulations.
The index is based on Ice-Core records of the climate state (see Sutter et al. 2019) and is used to scale forcing snapshots between the time slices.

For the atmosphere, the index is used to scale anomaly snapshots of the air temperature and precipitation.
The precipitation can also be scaled using either an uniform or spacial scaling factor (read from an input file).
In addition, the air temperature uses a cosinus approach using annual and mean summer temperatures for representing the yearly cycle (copied from YearlyCycle).

For the ocean, the climate index is implemented into PICO, and is used to scale anomaly snapshots of the potential temperature and salinity.