Add the configuration parameter "energy.model"
... and remove "energy.enabled" and "energy.temperature_based"

"energy.temperature_based" was used to determine if we should use "relaxed" checks for
temperate ice and (ages ago) to enable (if set) and disable (if not set) the isothermal
Glen flow law.

However, this relaxed check is appropriate when "energy.model" is "none" or
"cold" (exactly when "energy.temperature_based" was set) and the "isothermal_glen" flow
law can always be used in the current PISM version.

"energy.enabled" is false when "energy.model" is "none" and true otherwise.

The old command-line option "-energy ..." now maps to "energy.model" instead of this pair
of parameters.