Ignore prescribed retreat mask values that are very close to 1
For some reason YAC's interpolation may produce retreat mask values very slightly less
than one (around "1 - 1e-12") *in the interior* of the ice sheet even though the input
file contains *exact* ones everywhere in the interior.

This is not great, but I can't think of a way to do better.