Fix the statement about sub-shelf melt in the "-ocean th" section of the manual
I have no idea by we emphasized that sub-shelf melt is not fed back into the surrounding
ocean. PISM does not model sub-shelf-melt-ocean feedbacks *at all*, not matter which ocean
model component is selected.

It is OK to state that this ocean model component does not include a model of sub-shelf
circulation, though. This model is purely "local," i.e. one- (if you count the vertical
dimension) or zero-dimensional if you don't.