Add DimensionMetadata and use it in SpatialVariableMetadata
DimensionMetadata adds the length to VariableMetadata. This makes it easier to define a
dimension and set all the attributes: now we only need DimensionMetadata to do that.

SpatialVariableMetadata::x() (and y() and z()) return DimensionMetadata now. Now
SpatialVariableMetadata contains all the information needed to define required NetCDF
dimensions and coordinate variables (and set their attributes).

This change required an update to the SpatialVariableMetadata constructor: we need x and y
dimension lengths and grid spacing. I added two constructors: one extracts this
information from a Grid instance, the other one does not (to reduce coupling, if
necessary).