Avoid writing time-dependent 2D and 3D variables more than once
For example: before this change "pism -eisII A -Mx 5 -My 5 -energy enthalpy -o foo.nc"
saved the "enthalpy" variable *twice* because 1) it is a part of the state of the chosen
energy balance model and 2) it is a diagnostic listed in `output.sizes.medium`.

Now we keep track of time-dependent variables (we already did that for time-independent
ones) and avoid writing them more often than necessary.