Actually fix a bed deformation regression test (stop and re-start)
This completes the fix started in 79d8244e68b203d64520c66c1f0738ddbe73a466. 79d8244e fixed
an issue with a configuration parameter name (bed_deformation.update_interval) that was
out of date.

This commit fixes a different issue: stopping and re-starting PISM (in the old regression
test) affected time stepping, so the bed deformation model in the interrupted run "saw" a
different ice thickness history (which is used to compute the time average of the load on
the bed over bed_deformation.update_interval), so results were slightly different... but
not because of a failure to re-start the model properly.

The updated version of this test ensures that we stop and re-start without affecting time
step lengths.